requirements.txt
streamlit==1.41.1
anthropic==0.43.0
requests==2.32.3
beautifulsoup4==4.12.3
openpyxl==3.1.5
python-dotenv==1.0.1
lxml==5.3.0
